Changed journal mode of SQLiteDatabase class
In the class SQLiteDatabase, the journal mode has been changed from WAL (Write-Ahead Logging) to the default setting DELETE. The reason is that the WAL mode does not allow opening database files on remote drives. Note that this does not affect the settings store, which is steill opened in WAL mode, because it does not use the SQLiteDatabase class.