Fix bug with programm termination
Fixed a bug which prevented the program termination of HIT. It basically esclusively affected the automation mode, where the program termination immediatly follows after the last dataset has been processed (in interactive mode, there is essentially always some idle time after the last dataset is finished).
What happened was as follows: Whenever a dataset is finished, it generates a message with this information which is relayed to all listeners, synchronously and in the dataset processing thread. In the HRTImagingToolDialog class, this triggers OnSNPFusionEvent(), where a new message (to the dialog itself) was queued (asynchronously, to be executed in the main thread) and a new log line was created (synchronously, to be executed in the dataset processing thread), in this order. Therefore, after receiving the new message, the main thread could immediatly start with the termination process, which could effectively prevent the dataset processing thread to create the new log line (because the log window was already being destroyed).
The solution is simply to change the order of the two things happening in HRTImagingToolDialog::OnSNPFusionEvent(). Just to be on the safe side, a wait time of a second before destroying the dialog window has been added as well.