Minor changes for parameter dialog
The controls used for the new parameter (the acceptable residual error level) have been renamed to fit the naming scheme of the other controls.
The missing explanation for the new parameter has been added (some small changes were also necessary for the explanation of the correlation threshold for sub-image registration).