Default process parameters changed
The preset parameter sets have been adjusted to use the new automatic threshold detection. Note that it remains turned off in the base default settings; so if you start HIT for the first time in a specific user account, or if you remove the HIT parameter database file, the base default settings without automatic threhsold detection will be used.