Use Boost Format library for string formatting
Most string formatting is now implemented using Boost's Format library. This specifically excludes the HITgui project, where wxWidgets' wxString::Format() function is still used.
The reason for this change is the aim to remove the MFC (and potentially the Win32 API on the long run) from HIT. The last remaining occurrence of CString was in the CStringUtilities::Format() function. This is now gone, together with the entire CStringUtilities::Format() function; this may be considered collateral damage, but using the Boost Format API instead of calling the CStringUtilities::Format() function is not really a lot of hazzle.
In the wake of the removal of CStringUtilities::Format(), the string formatting capabilities of the CLog::Log() function were also removed. Again, the change does not create a lot of overhead code, and the Log() function can now actually use std::wstring arguments instead of LPCSTRs.