Direct support for boost format objects in CLog
The CLog::Log function is now overloaded to directly take a boost::wformat argument (instead of the std::wstring argument). This can save a little bit of code because the boost::wformat object does not need to be instantiated outside the Log() function, only to call the str() on it. The Log() function call can now be very similar to the removed variable argument list version.