Decreased dependence on Windows environment
Header file #include directives have been changed to reflect the correct file names (regarding capitalization). This was only necessary to change "StdAfx.h" to "stdafx.h".
Usage of __min() and __max() have been changed to std::min() and std::max().
MSVC allows file paths to be specified as std::wstring when opening streams, but this is not conforming to the standard specification. This has been changed to use std::filesystem::path instead of std::wstring (the path can be created from the wstring).
MSVC allows assignment of a std::wstring to a std::filesystem__path variable, but this is not conforming to the standard specification. This has been changed by using the path's wstring() method.
All calls to memcpy_s() are now restricted to Windows environments by "#ifdef _WIN32". In other environments, they have been replaced by the standard std::memcpy() function.