Decreased dependence on Windows environment
MSVC allows file paths to be specified as std::wstring when opening streams, but this is not conforming to the standard specification. This has been changed to use std::filesystem::path instead of std::wstring (the path can be created from the wstring).
MSVC allows assignment of a std::wstring to a std::filesystem__path variable, but this is not conforming to the standard specification. This has been changed by using the path's wstring() method.