Query environment variables on non-Windows systems
This is a first implementation of the CEnvironmentVariables class that supports (hopefully correctly) non-Windows environments. Ir compiles with gcc but has not been tested. The Windows implementation remains unchaged. The new code only uses standard library functions, though, so if it works correctly, also on Windows systems, it should entirely replace the old code.