Fix GUI log update
Appending a log line to the log text control is actually called from threads other than the GUI thread. This has been the case for a really long time, so it seems that it doesn't matter under Windows. But the correct solution is to do it asynchronously. It's easy with the CallAfter()-function of wxWidgets controls and C++ lambdas.