Update .gitignore
The folder "out" has been added to the list; this folder is not used in the specified setup, but it is the default output folder in CMake setups. So in order to avoid it being accidentally included in git if it is accidentally created in some environment, we want to explicitly ignore it.